+++ 扫雷图形版.vcxproj.filters
@@ -27,6 +27,9 @@
     <ClCompile Include="Minesweeper.cpp">
       <Filter>源文件</Filter>
     </ClCompile>
+    <ClCompile Include="Screen.cpp">
+      <Filter>源文件</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="acllib.h">
@@ -36,6 +39,12 @@
       <Filter>头文件</Filter>
     </ClInclude>
     <ClInclude Include="resource.h">
+      <Filter>头文件</Filter>
+    </ClInclude>
+    <ClInclude Include="Screen.h">
+      <Filter>头文件</Filter>
+    </ClInclude>
+    <ClInclude Include="setting.h">
       <Filter>头文件</Filter>
     </ClInclude>
   </ItemGroup>
